refactor(team generator): changed routes for better access
