+++ routes/overviews.js
@@ -10,6 +10,9 @@ const CRUD = require(`../controller/crud-operations`);
 
 const acronymGen = require(`../public/js/acronym-generator`);
 
+const team = require(`../controller/team-generator`);
+
+// Course overview
 router.get(`/courses`, ensureAuthenticated, (req, res) => {
     CRUD.findDocByQuery(schemas.User, `username`, req.params.username).then((userData) => {
 
@@ -66,7 +69,7 @@ router.get(`/courses`, ensureAuthenticated, (req, res) => {
     });
 });
 
-// Only students will navigate here
+//Class details directly after clicking on a course (only students will navigate here)
 router.get('/:course', ensureAuthenticated, (req, res) => {
     CRUD.findDocByQuery(schemas.User, 'username', req.params.username).then((userData) => {
 
@@ -109,7 +112,7 @@ router.get('/:course', ensureAuthenticated, (req, res) => {
 
                 allCourseClasses.forEach((courseClass) => {
                     if (allClasses.includes(courseClass._id)) {
-                        res.redirect(`${courseData.linkRef}/${courseClass.linkRef}_elective`);
+                        res.redirect(`${courseData.linkRef}/${courseClass.linkRef}`);
                     }
                 });
             });
@@ -117,7 +120,7 @@ router.get('/:course', ensureAuthenticated, (req, res) => {
     });
 });
 
-// Only teachers will navigate here
+//Get all classes of specific course (only teachers will navigate here)
 router.get(`/:course/classes`, ensureAuthenticated, (req, res) => {
     CRUD.findDocByQuery(schemas.Course, `linkRef`, req.params.course).then((paramCourse) => {
         CRUD.findDocByQuery(schemas.Teacher, `user`, req.user.id).then((user) => {
@@ -165,31 +168,42 @@ router.get(`/:course/classes`, ensureAuthenticated, (req, res) => {
     });
 });
 
-
+// Get class details (student overview)
 router.get(`/:course/:class`, ensureAuthenticated, (req, res) => {
-    const classType = req.params.class.substring(req.params.class.indexOf('_') + 1);
-    const classLink = req.params.class.split('_')[0];
 
-    CRUD.findDocByQuery(schemas.User, 'username', req.params.username).then((userData) => {
+    CRUD.findDocByQuery(schemas.Course, 'linkRef', req.params.course).then((courseData) => {
         let prevURL;
         let schema;
         let className;
         let formURL;
 
-        if (userData.type == 'student') {
-            if (req.params.course == 'class') {
+        const prevURLPath = new URL(req.headers.referer);
+        let url = prevURLPath.pathname.split('/');
+        const prevPage = url.pop();
+
+        if (req.user.type == 'student') {
+            if (prevPage == 'home') {
+                prevURL = `/${req.params.username}/${req.params.course}/${req.params.class}/home`;
+            } else if (req.params.course == 'class') {
                 prevURL = `/`;
             } else {
                 prevURL = `/${req.params.username}/courses`;
             }
 
-        } else if (userData.type == 'teacher') {
-            prevURL = `/${req.params.username}/${req.params.course}/classes`;
+        } else if (req.user.type == 'teacher') {
+
+            if (prevPage == 'home') {
+                prevURL = `/${req.params.username}/${req.params.course}/${req.params.class}/home`;
+            } else {
+                prevURL = `/${req.params.username}/${req.params.course}/classes`;
+            }
+
             className = 'overflow form--popup';
             formURL = `/teams/create`;
         }
 
-        if (classType != 'normal') {
+
+        if (courseData.type != 'normal') {
             schema = schemas.ElectiveClass;
         } else {
             schema = schemas.Class;
@@ -197,7 +211,7 @@ router.get(`/:course/:class`, ensureAuthenticated, (req, res) => {
 
         // insert user info based on id      
         schema.findOne({
-            'linkRef': classLink
+            'linkRef': req.params.class
         }).lean().populate({
             path: 'teachers',
             populate: {
@@ -216,7 +230,7 @@ router.get(`/:course/:class`, ensureAuthenticated, (req, res) => {
                 formURL: formURL,
                 isPopup: true,
                 studentData: classData.students,
-                userData: userData,
+                userData: req.user,
                 bannerTitle: classData.title,
                 bannerSubtitle: `${classData.students.length} studenten`,
                 linkRef: classData.linkRef,
@@ -227,10 +241,143 @@ router.get(`/:course/:class`, ensureAuthenticated, (req, res) => {
     });
 });
 
+
+// Team form page (teacher can only get here when JS is disabled)
+router.get(`/:course/:class/team-form`, ensureAuthenticated, (req, res) => {
+    let schema;
+
+    CRUD.findDocByQuery(schemas.Course, 'linkRef', req.params.course).then((courseData) => {
+
+        if (courseData.type != 'normal') {
+            schema = schemas.ElectiveClass;
+        } else {
+            schema = schemas.Class;
+        }
+
+        CRUD.findDocByQuery(schema, 'linkRef', req.params.class).then((classObject) => {
+
+            res.render('team-generation', {
+                bannerTitle: classObject.title,
+                bannerSubtitle: `${classObject.students.length} studenten`,
+                linkRef: classObject.linkRef,
+                isPopup: false,
+                formURL: `/${req.params.username}/${req.params.course}/${req.params.class}/teams/create`,
+                prevURL: `/${req.params.username}/${req.params.course}/${req.params.class}`,
+                classTeams: classObject.teams
+            });
+        });
+    });
+});
+
+router.post('/:course/:class/teams/create', (req, res) => {
+    const teamSize = req.body.teamSize;
+
+    let allStudentObjects = [];
+    let allTeams = [];
+
+    CRUD.findDocByQuery(schemas.Course, 'linkRef', req.params.course).then((courseData) => {
+        let classSchema;
+
+        if (courseData.type != 'normal') {
+            classSchema = schemas.ElectiveClass;
+        } else {
+            classSchema = schemas.Class;
+        }
+
+        // reset
+        schemas.Team.remove((err) => {
+            if (err) Promise.reject(err);
+            console.log(`deleted`);
+        });
+
+        classSchema.updateMany({
+            linkRef: req.params.class
+        }, {
+            $set: {
+                teams: []
+            }
+        }, (err, affected) => {
+            console.log(`affected`, affected);
+        });
+
+        schemas.Student.updateMany({}, {
+            $set: {
+                teams: []
+            }
+        }, (err, affected) => {
+            console.log(`affected`, affected);
+        });
+
+        classSchema.findOne({
+            'linkRef': req.params.class
+        }).lean().populate({
+            path: 'students',
+            populate: {
+                path: 'user'
+            }
+        }).exec((err, classData) => {
+            if (err) Promise.reject(err);
+
+            allStudentObjects = classData.students;
+
+            team.generate(allStudentObjects, teamSize).then((generatedTeams) => {
+                generatedTeams.forEach((team) => {
+
+                    CRUD.createDoc(schemas.Team, {
+                        name: team.name,
+                        number: team.number,
+                        students: team.students,
+                        class: classData.id,
+                        course: courseData.id
+                    }).then((doc) => {
+
+                        // add team id to class teams
+                        CRUD.addIdReferenceToDoc(classSchema, classData._id, 'teams', doc._id);
+
+                        doc.students.forEach((object) => {
+                            // add team id to teams of each student
+                            CRUD.addIdReferenceToDoc(schemas.Student, object.student._id, 'teams', doc._id);
+                        });
+
+                        schemas.Team.findById(doc.id).lean().populate({
+                            path: `students.student`,
+                            select: `user`,
+                            populate: {
+                                path: `user`,
+                                select: `name profile_pic`
+                            }
+                        }).populate('students.cmd_skill').exec((err, teamData) => {
+                            if (err) Promise.reject(err);
+
+                            allTeams.push(teamData);
+
+                            if (allTeams.length === generatedTeams.length) {
+                                allTeams.forEach((team) => {
+                                    team.students.forEach(member => {
+                                        console.log(member.cmd_skill);
+                                    });
+                                });
+                                res.render('team-details', {
+                                    allTeams: allTeams.sort((a, b) => parseFloat(a.number) - parseFloat(b.number)),
+                                    userData: null,
+                                    courseData: null,
+                                    memberView: false
+                                });
+                            }
+                        });
+                    });
+                });
+            });
+        });
+    });
+});
+
+
+
+// Class home page (only visible if class has teams already)
 router.get(`/:course/:class/home`, ensureAuthenticated, (req, res) => {
     let prevURL;
     let schema;
-    let linkRef;
 
     if (req.user.type == 'student') {
         prevURL = `/${req.params.username}/courses`;
@@ -242,10 +389,8 @@ router.get(`/:course/:class/home`, ensureAuthenticated, (req, res) => {
 
         if (courseData.type != 'normal') {
             schema = schemas.ElectiveClass;
-            linkRef = `${courseData.linkRef}/${req.params.class}_elective`;
         } else {
             schema = schemas.Class;
-            linkRef = `${courseData.linkRef}/${req.params.class}_normal`;
         }
 
         CRUD.findDocByQuery(schema, `linkRef`, req.params.class).then((classObject) => {
@@ -256,48 +401,44 @@ router.get(`/:course/:class/home`, ensureAuthenticated, (req, res) => {
                 prevURL: prevURL,
                 bannerTitle: classObject.title,
                 bannerSubtitle: `${classObject.students.length} studenten`,
-                linkRef: linkRef
+                linkRef: `${courseData.linkRef}/${req.params.class}`
             });
         });
     });
 });
 
+// Get all teams of specific course class
 router.get(`/:course/:class/teams`, ensureAuthenticated, (req, res) => {
     let schema;
-    let linkRef;
 
-    const classType = req.params.class.substring(req.params.class.indexOf('_') + 1);
-    const classLink = req.params.class.split('_')[0];
+    CRUD.findDocByQuery(schemas.Course, 'linkRef', req.params.course).then((courseData) => {
+        if (courseData.type != 'normal') {
+            schema = schemas.ElectiveClass;
+        } else {
+            schema = schemas.Class;
+        }
 
-    if (classType != 'normal') {
-        schema = schemas.ElectiveClass;
-        linkRef = `${req.params.course}/${req.params.class}_elective`;
-    } else {
-        schema = schemas.Class;
-        linkRef = `${req.params.course}/${req.params.class}_normal`;
-    }
+        schema.findOne({
+            'linkRef': req.params.class
+        }).lean().populate({
+            path: `teams`,
+            select: `name students number`
+        }).exec((err, classData) => {
+            if (err) Promise.reject(err);
+            console.log(classData);
 
-    schema.findOne({
-        'linkRef': classLink
-    }).lean().populate({
-        path: `teams`,
-        select: `name students number`
-    }).exec((err, classData) => {
-        if (err) Promise.reject(err);
-        console.log(classData);
-
-        res.render(`team-overview`, {
-            prevURL: `/${req.params.username}/${req.params.course}/${req.params.class}/home`,
-            formURL: `/teams/create`,
-            userData: null,
-            isPopup: true,
-            teamData: classData.teams.sort((a, b) => parseFloat(a.number) - parseFloat(b.number)),
-            bannerTitle: classData.title,
-            bannerSubtitle: `Team overzicht`,
-            linkRef: linkRef
+            res.render(`team-overview`, {
+                prevURL: `/${req.params.username}/${req.params.course}/${req.params.class}/home`,
+                formURL: `/${req.params.username}/${req.params.course}/${req.params.class}/teams/create`,
+                userData: null,
+                isPopup: true,
+                teamData: classData.teams.sort((a, b) => parseFloat(a.number) - parseFloat(b.number)),
+                bannerTitle: classData.title,
+                bannerSubtitle: `Team overzicht`,
+                linkRef: `${req.params.course}/${req.params.class}`
+            });
         });
     });
-
 });
 
 module.exports = router;
