+++ views/class-details.ejs
@@ -8,16 +8,16 @@
             <% studentData.forEach(item => { %>
                 <li>
                     <a class="box-shadow" href="#">
-                        <img src="../../public/<%= item.user.profile_pic %>" alt="Profielfoto van <%= item.user.name %>">
+                        <img src="/public/<%= item.user.profile_pic %>" alt="Profielfoto van <%= item.user.name %>">
                     </a>
                 </li>
             <% }) %>
         </ul>
 
         <div class="btn--container">
-            <a class="btn--small btn--small-arrow" href="<%= prevURL %>"><img src="../../public/images/icons/arrow-icon.svg" alt="Back icon"></a>
+            <a class="btn--small btn--small-arrow" href="<%= prevURL %>"><img src="/public/images/icons/arrow-icon.svg" alt="Back icon"></a>
             <% if (userData.type == 'teacher' && !classTeams.length) { %>
-                <a class="btn form-trigger" href="/teams">Teams aanmaken</a>
+                <a class="btn form-trigger" href="teams">Teams aanmaken</a>
             <% } %>
         </div>
 
@@ -28,5 +28,5 @@
 </body>
 
 <% if (classTeams != undefined && !classTeams.length) { %>
-     <script async src="../../public/js/form-script.js"></script>
+     <script async src="/public/js/form-script.js"></script>
 <% } %>
