+++ views/team-details.ejs
@@ -39,7 +39,7 @@
         </ul>
 
         <div class="btn--container">
-            <a class="btn--small btn--small-arrow" href="<%= prevURL %>"><img src="../public/images/icons/arrow-icon.svg" alt="Back icon"></a>
+            <a class="btn--small btn--small-arrow" href="<%= prevURL %>"><img src="/public/images/icons/arrow-icon.svg" alt="Back icon"></a>
         </div>
 
         <% } else { %>
@@ -78,7 +78,7 @@
     </main>
 
     <% if (!memberView) { %>
-        <script async src="../public/js/team-details-script.js"></script>
+        <script async src="/public/js/team-details-script.js"></script>
     <% } %>
 
 </body>
