+++ views/team-overview.ejs
@@ -29,8 +29,8 @@
         </ol>
 
         <div class="btn--container">
-            <a class="btn--small btn--small-arrow" href="#"><img src="/public/images/icons/arrow-icon.svg" alt="Back icon"></a>
-            <a class="btn form-trigger" href="/team-generation/form">Nieuwe teams maken</a>
+            <a class="btn--small btn--small-arrow" href="<%= prevURL %>"><img src="/public/images/icons/arrow-icon.svg" alt="Back icon"></a>
+            <a class="btn form-trigger" href="<%= linkRef %>/team-form">Nieuwe teams maken</a>
         </div>
 
         <%- include ('./partials/form.ejs') -%>
